+++ changelog.md
@@ -10,6 +10,11 @@
 - Ship spa, ita, fra, ron and por OCR languages data in docker image [Issue#586](https://github.com/ciur/papermerge/issues/586)
 - Fix create_token.sh throws an error [Issue#314](https://github.com/papermerge/papermerge-core/issues/314)
 
+### Adds
+
+- [UI] Context Missing "Extract Pages" entry [Issue#564](https://github.com/ciur/papermerge/issues/564)
+
+
 ## [3.0.2] - 2024-01-21
 
 ### Fixes